Changes to collectd.c
